[WIP][FEATURE] Fix problem with cache being cleared multiple time
* Applied manual to FVM site !!! *

This commit is a proof-of-concept and is to be used as a basis for further
discussion.

When pages or records in general are changed using the DataHandler,
the pageCache is cleared for each page multiple times. To fix
this, the DataHandler will collect which pages needs clearing, and
wait with the actual clearing until all operations in cmdap og dataMap
has bee processed. This way, the cache is only cleared once for each
page.

The commit also introduced a new configuration, so not every cache
is asked to clear for tags pageId_<PageUid>. Default is pages, pageSection
and cache_hash.

Change-Id: If1aa389991b801413eafa38d7d140f932bb321f2
Refs: #52235