[FEATURE][BUGFIX] Make Apc and Memecache clearing much more efficient
When calling flush, both acp and memcache call find by identifer to
find all tags belonging to the current site. Since we know that FVM
is the only one using the APC and Memcache caces, we have modified it
to simply call the relevant Flush commands.

This patch can be removed when we have changed to Redis backend which
does not have this artifact.